Check to make sure YAML doesn't already exist before copying

+++ lib/tasks/test.rake
@@ -1,7 +1,10 @@
 namespace :test do
   desc "Configure the test environment"
   task :prepare => :environment do
-    cp "#{Rails.root}/config/status_exchange.yml.example", "#{Rails.root}/config/status_exchange.yml"
+    unless File.exists? "#{Rails.root}/config/status_exchange.yml"
+      cp "#{Rails.root}/config/status_exchange.yml.example", "#{Rails.root}/config/status_exchange.yml"
+    end
+
     Rake::Task['test:prepare'].invoke
   end
 end
